pyproject.toml
[project]
name = "ai-group-chat"
version = "0.1.0"
description = "ai群聊项目～启动！"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.27.0",
    "pyautogen>=0.2.0",
    "autogen-ext[openai]>=0.4.0",
    "openai>=1.50.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "uvicorn>=0.40.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "tiktoken>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_group_chat"]

[project.scripts]
ai-chat = "ai_group_chat.main:app"
